modules/dataset_wrangling/pyproject.toml
[tool.poetry]
name = "dataset-wrangling"
version = "0.1.0"
description = ""
authors = ["Pranav Suryadevara <suryadevarapranav28@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pydantic = "^2.9.2"
dspy-ai = "^2.5.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
